+++ mayan/apps/document_states/views/workflow_proxy_views.py
@@ -1,7 +1,6 @@
 from django.template import RequestContext
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.documents.models import Document
 from mayan.apps.documents.views.document_views import DocumentListView
 from mayan.apps.views.generics import SingleObjectListView
 from mayan.apps.views.mixins import ExternalObjectViewMixin
@@ -20,9 +19,7 @@ class WorkflowRuntimeProxyDocumentListView(
     external_object_pk_url_kwarg = 'workflow_runtime_proxy_id'
 
     def get_document_queryset(self):
-        return Document.valid.filter(
-            workflows__workflow=self.external_object
-        )
+        return self.external_object.get_documents()
 
     def get_extra_context(self):
         context = super().get_extra_context()