+++ mayan/apps/mirroring/caches.py
@@ -8,34 +8,37 @@ from .settings import (
 )
 
 
-class IndexFilesystemCache:
+class MirrorFilesystemCache:
     @staticmethod
     def get_key_hash(key):
         return hashlib.sha256(force_bytes(s=key)).hexdigest()
 
     @staticmethod
     def get_document_key(document):
-        return IndexFilesystemCache.get_key_hash(
+        return MirrorFilesystemCache.get_key_hash(
             key='document_pk_{}'.format(document.pk)
         )
 
     @staticmethod
     def get_node_key(node):
-        return IndexFilesystemCache.get_key_hash(
+        return MirrorFilesystemCache.get_key_hash(
             key='node_pk_{}'.format(node.pk)
         )
 
     @staticmethod
     def get_path_key(path):
-        return IndexFilesystemCache.get_key_hash(
+        return MirrorFilesystemCache.get_key_hash(
             key='path_{}'.format(path)
         )
 
     def __init__(self, name='default'):
         self.cache = caches[name]
 
+    def clear_all(self):
+        self.cache.clear()
+
     def clear_node(self, node):
-        node_key = IndexFilesystemCache.get_node_key(node=node)
+        node_key = MirrorFilesystemCache.get_node_key(node=node)
         path_cache = self.cache.get(key=node_key)
         if path_cache:
             path = path_cache.get('path')
@@ -45,7 +48,7 @@ class IndexFilesystemCache:
         self.cache.delete(key=node_key)
 
     def clear_document(self, document):
-        document_key = IndexFilesystemCache.get_document_key(document=document)
+        document_key = MirrorFilesystemCache.get_document_key(document=document)
         path_cache = self.cache.get(key=document_key)
         if path_cache:
             path = path_cache.get('path')
@@ -56,38 +59,36 @@ class IndexFilesystemCache:
 
     def clean_path(self, path):
         self.cache.delete(
-            key=IndexFilesystemCache.get_path_key(path=path)
+            key=MirrorFilesystemCache.get_path_key(path=path)
         )
 
     def get_path(self, path):
         return self.cache.get(
-            key=IndexFilesystemCache.get_path_key(path=path)
+            key=MirrorFilesystemCache.get_path_key(path=path)
         )
 
     def set_path(self, path, document=None, node=None):
         # Must provide a document_pk or a node_pk
-        # Not both
-        # Providing both is also not correct
-
+        # not both.
         if document:
             self.cache.set(
-                key=IndexFilesystemCache.get_path_key(path=path),
+                key=MirrorFilesystemCache.get_path_key(path=path),
                 value={'document_pk': document.pk},
                 timeout=setting_document_lookup_cache_timeout.value
             )
             self.cache.set(
-                key=IndexFilesystemCache.get_document_key(document=document),
+                key=MirrorFilesystemCache.get_document_key(document=document),
                 value={'path': path},
                 timeout=setting_document_lookup_cache_timeout.value
             )
         elif node:
             self.cache.set(
-                key=IndexFilesystemCache.get_path_key(path=path),
+                key=MirrorFilesystemCache.get_path_key(path=path),
                 value={'node_pk': node.pk},
                 timeout=setting_node_lookup_cache_timeout.value
             )
             self.cache.set(
-                key=IndexFilesystemCache.get_node_key(node=node),
+                key=MirrorFilesystemCache.get_node_key(node=node),
                 value={'path': path},
                 timeout=setting_node_lookup_cache_timeout.value
             )