+++ mayan/apps/mirroring/tests/literals.py
@@ -1,9 +1,19 @@
 # -*- coding: utf-8 -*-
 
+TEST_CABINET_LABEL = 'level_1'
+TEST_CABINET_LABEL_INVALID = 'level/1'
+TEST_CABINET_LABEL_MULTILINE = 'first\r\nsecond\r\nthird'
+TEST_CABINET_LABEL_MULTILINE_EXPECTED = 'first second third'
+TEST_CABINET_LABEL_MULTILINE_2 = '\r\n\r\nfirst\r\nsecond\r\nthird\r\n'
+TEST_CABINET_LABEL_MULTILINE_2_EXPECTED = 'first second third'
+
 TEST_CACHE_KEY_BAD_CHARACTERS = ' \r\n!@#$%^&*()+_{}|:"<>?-=[];\',./'
+
 TEST_DOCUMENT_PK = 99
+
 TEST_KEY_UNICODE = 'áéíóúüäåéë¹²³¤'
 TEST_KEY_UNICODE_HASH = 'ba418878794230c3f4308e66c70db31dd83f1def4d9381f379c50f42eb88989c'
+
 TEST_NODE_EXPRESSION = 'level_1'
 TEST_NODE_EXPRESSION_INVALID = 'level/1'
 TEST_NODE_EXPRESSION_MULTILINE = 'first\r\nsecond\r\nthird'